tests/conftest.py에서 모든 모델 행을 만듬
모델이 서로 참조 중이고 똑같은 모델을 만느는 상황이
발생해서 하나의 파일에서 만듬
